backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
SQLAlchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
aioredis==2.0.1
httpx==0.27.2
aiohttp==3.9.0
python-dotenv==1.0.0
